+++ kgcnn/training/schedule.py
@@ -1,8 +1,10 @@
 import numpy as np
 import tensorflow as tf
 
+ks = tf.keras
 
-@tf.keras.utils.register_keras_serializable(package='kgcnn', name='LinearWarmupExponentialDecay')
+
+@ks.utils.register_keras_serializable(package='kgcnn', name='LinearWarmupExponentialDecay')
 class LinearWarmupExponentialDecay(tf.optimizers.schedules.LearningRateSchedule):
     r"""This schedule combines a linear warmup with an exponential decay.
     Combines :obj:` tf.optimizers.schedules.PolynomialDecay` with an actual increase during warmup
@@ -57,6 +59,34 @@ class LinearWarmupExponentialDecay(tf.optimizers.schedules.LearningRateSchedule)
     @initial_learning_rate.setter
     def initial_learning_rate(self, value):
         self.decay.initial_learning_rate = value
+
+    def get_config(self):
+        """Get config for this class."""
+        config = {}
+        config.update(self._input_config_settings)
+        return config
+
+
+@ks.utils.register_keras_serializable(package='kgcnn', name='KerasPolynomialDecaySchedule')
+class KerasPolynomialDecaySchedule(ks.optimizers.schedules.PolynomialDecay):
+    r"""This schedule extends :obj:` tf.optimizers.schedules.PolynomialDecay` ."""
+
+    def __init__(self, dataset_size: int, batch_size: int, epochs: int, lr_start: float = 0.0005,
+                 lr_stop: float = 1e-5):
+        """Initialize class.
+
+        Args:
+            dataset_size (int): Size of the dataset.
+            batch_size (int): Batch size for training.
+            epochs (int): Total epochs to run schedule on.
+            lr_start (int): Learning rate at the start.
+            lr_stop (int): Final learning rate.
+        """
+        self._input_config_settings = {"lr_start": lr_start, "lr_stop": lr_stop,
+                                       "epochs": epochs, "batch_size": batch_size, "dataset_size": dataset_size}
+        steps_per_epoch = dataset_size / batch_size
+        num_steps = epochs * steps_per_epoch
+        super().__init__(initial_learning_rate=lr_start, decay_steps=num_steps, end_learning_rate=lr_stop)
 
     def get_config(self):
         """Get config for this class."""