+++ changelog.md
@@ -3,7 +3,7 @@ v3.0.1
 * Removed deprecated molecules.
 * Fix error in ``kgcnn.data.transform.scaler.serial``
 * Fix error in ``QMDataset`` for molecular features.
-* 
+* Fix error in ``kgcnn.layers.conv.GraphSageNodeLayer`` .
 
 
 v3.0.0